Ticket#2021041010000013: Terminate rec on exit.
When started by daemon(8), daemon forwards stdout/stderr to its stdout.
During "service stop", i.e. when daemon processes the SIGTERM it gets
from the rc script, it closes its pipe to the recorder manager process,
and the attempt to announce on stderr that rec is about to be killed
fails via SIGPIPE.

Fixed by removing that print().